% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacent.R
\name{adjacent_join}
\alias{adjacent_join}
\title{Given a table representing a grid, join to adjacent cells in the grid}
\usage{
adjacent_join(x, y = x, diagonal = FALSE, suffix = c("", "2"))
}
\arguments{
\item{x}{A table with columns for row and col}

\item{y}{A second table with columns row and col. By default, does
a self-join}

\item{diagonal}{Whether diagonal cells are counted as adjacent}

\item{suffix}{By default, "" and "2"}
}
\description{
Given a table representing a grid, join to adjacent cells in the grid
}
