% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacent.R
\name{adjacent_nest}
\alias{adjacent_nest}
\title{Nest a column with adjacent values}
\usage{
adjacent_nest(x, y = x, diagonal = FALSE)
}
\arguments{
\item{x}{A table with columns for row and col}

\item{y}{A second table with columns row and col. By default, does
a self-join}

\item{diagonal}{Whether diagonal cells are counted as adjacent}
}
\description{
Nest a column with adjacent values
}
\examples{

library(purrr)
library(dplyr)
grid11 <- advent_input(11) \%>\%
  grid_tidy(x)

# One "flash": any cells with a value above 9 cause neighbors to increase
grid11 \%>\%
  mutate(value = value + 1) \%>\%
  adjacent_nest(diagonal = TRUE) \%>\%
  mutate(value = value + map_dbl(adjacent_value, ~ sum(. > 9)))

}
