% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{advent_input}
\alias{advent_input}
\title{Read in an Advent of Code input as a one-column tibble}
\usage{
advent_input(day = lubridate::day(Sys.Date()), year = 2021, parse = FALSE)
}
\arguments{
\item{day}{Day within December}

\item{year}{By default, 2021}

\item{parse}{Whether to parse the input using \code{parse} (e.g. as a number)}
}
\description{
Read in an Advent of Code input as a one-column tibble
}
