% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\docType{data}
\name{HEX_DIRECTIONS}
\alias{HEX_DIRECTIONS}
\title{Hexagonal directions in a 3d grid}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 5 columns.
}
\usage{
HEX_DIRECTIONS
}
\description{
Either oriented facing north/south or east/west
}
\seealso{
\url{https://www.redblobgames.com/grids/hexagons/}
}
\keyword{datasets}
