% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_txt}
\alias{parse_txt}
\title{Parse an input that may be a number}
\usage{
parse_txt(s, ...)
}
\arguments{
\item{s}{A character vector}

\item{...}{Extra arguments passed on to parse_guess}
}
\description{
Uses \code{\link[readr]{parse_guess}}, but ignores commas (commas in
Advent of Code are almost always delimiters)
}
