% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid_tidy}
\alias{grid_tidy}
\alias{grid_matrix}
\alias{grid_graph}
\title{Parse a table with a string column that represents rows in a grid}
\usage{
grid_tidy(d, var, sep = "", parse = TRUE)

grid_matrix(d, var, sep = "", parse = TRUE)

grid_graph(d, ...)
}
\arguments{
\item{d}{A table}

\item{var}{The name of a string column}

\item{sep}{A string separator to use for splitting columns}

\item{parse}{Whether to parse the input by guessing its format}

\item{...}{Any of the above arguments}
}
\description{
Many Advent of Code challenges involve parsing a grid where each row
of the input represents one row of the grid. These turn such inputs into
a tidy table, a matrix, or a tidygraph object.
}
